fix notification crash issue with android module
