+++ .gitignore
@@ -39,3 +39,4 @@ service_identity/local.properties
 /client_dashboard/.idea/shelf/Uncommitted_changes_before_Checkout_at_8_14_2023_5_20_PM_[Changes]/shelved.patch
 /client_dashboard/.idea/shelf/Uncommitted_changes_before_Checkout_at_8_14_2023_5_20_PM_[Changes]1/shelved.patch
 /client_dashboard/.idea/shelf/Uncommitted_changes_before_Checkout_at_8_14_2023_5_18_PM_[Changes]/shelved.patch
+/client_end_user/androidApp/google-services.json
