+++ client_end_user/androidApp/build.gradle.kts
@@ -2,6 +2,7 @@ plugins {
     kotlin("multiplatform")
     id("com.android.application")
     id("org.jetbrains.compose")
+    id("com.google.gms.google-services") version "4.4.0"
 }
 
 kotlin {
@@ -39,7 +40,7 @@ android {
         jvmToolchain(libs.versions.jvmToolchain.get().toInt())
     }
 }
-dependencies{
+dependencies {
     implementation("androidx.core:core-splashscreen:1.0.1")
     implementation("com.google.firebase:firebase-common-ktx:20.4.2")
     implementation(libs.firebase.messaging.ktx)
