--- client_end_user/androidApp/src/androidMain/kotlin/com/beepbeep/notification/FirebaseMessagingImpl.kt
@@ -1,19 +1,0 @@
-package com.beepbeep.notification
-
-import com.google.firebase.messaging.FirebaseMessaging
-import data.gateway.service.IFireBaseMessageService
-import kotlin.coroutines.resume
-import kotlin.coroutines.resumeWithException
-import kotlin.coroutines.suspendCoroutine
-
-class FirebaseMessagingImpl() : IFireBaseMessageService {
-    override suspend fun getDeviceToken(): String {
-        return suspendCoroutine {
-            FirebaseMessaging.getInstance().token.addOnSuccessListener { token ->
-                it.resume(token)
-            }.addOnFailureListener { exception ->
-                it.resumeWithException(exception)
-            }
-        }
-    }
-}
