+++ client_end_user/shared/src/commonMain/kotlin/data/gateway/remote/UserGateway.kt
@@ -30,7 +30,7 @@ import kotlinx.serialization.json.Json
 
 class UserGateway(
     client: HttpClient,
-    private val firebaseService: IFireBaseMessageService,
+    private val firebaseService: IFireBaseMessageService
 ) : BaseGateway(client), IUserGateway {
 
     @OptIn(InternalAPI::class)
