+++ gradle/libs.versions.toml
@@ -39,6 +39,7 @@ androidx-material3-version = "1.1.1"
 koin-compose-version = "1.0.4"
 moko-geo-version = "0.6.0"
 paging-compose-version="3.3.0-alpha02-0.4.0"
+google-services-version = "4.4.0"
 
 [libraries]
 androidx-core-ktx = { module = "androidx.core:core-ktx", version.ref = "androidx-core" }
@@ -112,4 +113,4 @@ google-dagger-hilt = { id = "com.google.dagger.hilt.android", version.ref = "hil
 kotlinKsp = { id = "com.google.devtools.ksp", version.ref = "ksp" }
 realm = { id = "io.realm.kotlin", version.ref = "realm-version" }
 kotlinx-serialization = { id = "org.jetbrains.kotlin.plugin.serialization", version.ref = "kotlinx-serialization-version" }
-
+google-services = { id = "com.google.gms.google-services",  version.ref = "google-services-version" }
